Create a node

Allows the user to define a new node.

Usage

node(cond, goT = "END", goF = "END", doT = "", doF = "")

Arguments

cond

A condition defined by has_vec, has_val, has_each_val or did_call.

goT

The node to which the evaluator moves if the condition in cond is true.

goF

The node to which the evaluator moves if the condition in cond is false.

doT

The message to display if the condition in cond is true.

doF

The message to display if the condition in cond is true.

See also

Examples

node(did_call('plot','base'),goT=S2,goF="END",doT="message displayed if true",doF="message displayed if false")
#>                  condition goT goF                       doT
#> 1 did_call("plot", "base")  S2 END message displayed if true
#>                          doF
#> 1 message displayed if false