Function for running a set of scripts Use with caution! If you run this script, all submitted R scripts will be executed automatically. Be sure that you trust these scripts - they could severely harm your computer.

Function for running a set of scripts Use with caution! If you run this script, all submitted R scripts will be executed automatically. Be sure that you trust these scripts - they could severely harm your computer.

Usage

runBatch(
  dirname,
  assignment,
  pattern = ".*\\.R$",
  rec = FALSE,
  overridewarning = FALSE
)

Arguments

dirname

Character, the directory that contains the scripts for analysing.

assignment

An assignment, prepared with prepareAssignment.

pattern

The pattern that the R scripts adhere to. By default all files ending in .R are found.

rec

Whether or not to find files recursively, i.e. in subdirectories too.

overridewarnings

Whether or not to show a warning before running the function.

Value

A list containing detailed output on each of the R scripts in the submitted directory. The output can be passed to analyseBatch, extractScripts or extractFeedback for further processing.